Revert "Rename htmlPlugin to fixLokiRefs"
This reverts commit e3629a8559c549981a30285085e1244418843aa3.
