+++ htmlPlugin.js
